pyproject.toml
[project]
name = "wishlist"
version = "0.1.0"
description = ""
authors = [
    {name = "Evgeny1337",email = "genya-sozykin@mal.ru"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (>=4.2)",
    "django-ninja (>=1.6.2,<2.0.0)",
    "pydantic (>=2.11)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-django (>=4.12.0,<5.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]
